zynq: Re-implement mkits.sh and use it
The new mkits.sh is a complete new implementation to support:
- multi-configuration (upto 10 with -c option). First -c used defines
the default conf used.
- multi-image support - multiple kernel/fdt/ramdsik image (upto 50 images)
- per image configuration: Tools now reuse the option for each images
  - hash algorithm and generated required subnodes
  - compression
  - signature and generated required subnodes (not tested)

This allows user to generate a simple its that only has one configuration
that only boot with 1 kernel + 1 fdt to 4 different boot configuration
with different image combination. For an example:
creating a its with 2 kernel, 1 dtb, 2 rootfs images and has the following
configurations:
	1: kernel0 + dtb0 + rootfs0 (default boot configuration)
	2. kernel0 + dtb0 + rootfs1
	3. kernel1 + dtb0 + rootfs0
	4. kernel1 + dtb0 + rootfs1
	5. kernel0 + dtb1
where kernel1 and rootfs0 is gzip compressed.

The command to generate the its is shown as follows:
	$ ./mkits.sh -A arm -v 4.4 -k kernel0 -C none -c 1 -c 2 -c5 \
	  -k kernel1 -C gzip -c 3 -c 4 -h crc32 \
	  -d dtb0 -c 1 -c 2 -c 3 -c 4 -h sha1 -h crc32\
	  -d dtb1 -c 5 -h sha1 \
	  -r rootfs0 -C gzip -c 1 -c 2 \
	  -r rootfs1 -C none -c3 -c4

First -c option used defines the default boot configuration.

The key benefit is to have more flexible ITS generation. A typical example,
is to create a fitImage with 1 kernel with multiple dtb to provide different
features. E.g. system boot with ramdisk or sdroot or nfsroot and etc.

Signed-off-by: Jason Wu <jason.wu.misc@gmail.com>